Moves `RemoteFeedItem` to its own file
